requirements.txt
# 高新医疗UDID查询系统 - Python依赖
# 部署日期: 2026-02-02
# Python版本要求: >=3.9

# ==========================================
# Web框架 (Web Framework)
# ==========================================
flask>=2.3.0,<3.0.0
flask-cors>=4.0.0,<5.0.0

# ==========================================
# 生产服务器 (Production Server)
# ==========================================
gunicorn>=21.0.0,<23.0.0

# ==========================================
# HTTP请求 (HTTP Requests)
# ==========================================
requests>=2.28.0,<3.0.0

# ==========================================
# 数据处理 (Data Processing)
# ==========================================
pandas>=1.5.0,<2.0.0
numpy>=1.24.0,<2.0.0
lxml>=4.9.0,<5.0.0

# ==========================================
# 中文文本处理 (Chinese Text Processing)
# ==========================================
jieba>=0.42.0

# ==========================================
# XML安全解析 (Secure XML Parsing)
# ==========================================
defusedxml>=0.7.0

# ==========================================
# 向量检索 (Vector Search)
# ==========================================
faiss-cpu>=1.7.4

# ==========================================
# 环境变量管理 (Environment Management)
# ==========================================
python-dotenv>=1.0.0

# ==========================================
# PostgreSQL 驱动 (PostgreSQL Driver)
# ==========================================
psycopg2-binary>=2.9.0

# ==========================================
# 可选依赖 (Optional Dependencies)
# ==========================================
# 如需GPU加速向量检索，替换faiss-cpu为:
# faiss-gpu>=1.7.4

# 性能监控 (可选)
# psutil>=5.9.0

# 缓存支持 (可选)
# redis>=4.5.0
